Creacion paquete Utilidades

+++ src/proyectosimulacion/controladores/utilidades/Util.java
@@ -1,13 +1,17 @@
 
 package proyectosimulacion.controladores.utilidades;
 
+import java.util.Scanner;
+
 
 public class Util {
     
     //Metodo para generar numeros pseudoAleatorios
-    private double GenerarAleatorio(){
+    private long GenerarAleatorio(){
         long random = 0;
-        
+                
         return random;
+        
+        
     }
 }
